Support Bookworm, exclude 64-bit and Pi 5 (#78)
* Bookworm 32-Bit Uplift (#77)

* Bookworm 32 (#76)

* Update libs

* Update README

* Fix for #58

* Static link gcc libs

* Move apt packages up front

* Restrict to Bullseye (or newer)

* Update install.sh (#66) (#74)

* Fixed a typo in the 'Important Note' section - Co-authored-by: Craig Ceremuga <c.ceremuga@gmail.com>

* Offline freq list (#75)

* Allow TX with no ini

* Update .gitignore

* Freeze Python env for scripts

* Rename and lint *.py

* Fully define development venv

* Add a link to Developer Notes

* Shellcheck *.sh

* Smarter directory work

* Better script logic

* Update file header strings

* Update dependencies

* Fix devel installers

* Prevent using a Pi 5

+++ scripts/copysite.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
 
+# Copyright (C) 2023-2024 Lee C. Bussy (@LBussy)
+# Created for WsprryPi version 1.2.1 [devel).
+
+# Get repo root
+repo_root=$(git rev-parse --show-toplevel 2>/dev/null)
+if [ -z "$repo_root" ]; then
+    echo "Not in a Git repository."
+    exit
+fi
+
 sudo rm -fr /var/www/html/wspr
 sudo mkdir /var/www/html/wspr
-sudo cp -R /home/pi/WsprryPi/data/* /var/www/html/wspr/
+sudo cp -R "$repo_root"/data/* /var/www/html/wspr/
 sudo ln -sf /usr/local/etc/wspr.ini /var/www/html/wspr/wspr.ini
 sudo chown -R www-data:www-data /var/www/html/wspr/
+echo "Website copied."
